fix docs: interactive TOC dependence on the page loading order
