+++ scripts/docs/doxygen-awesome-css/README.md
@@ -2,4 +2,7 @@
 
 Doxygen Awesome repository: [link](https://github.com/jothepro/doxygen-awesome-css)
 
-Version 2.3.3 (commit 28ed396de19cd3d803bcb483dceefdb6d03b1b2b) without changes
+Version 2.3.3 (commit 28ed396de19cd3d803bcb483dceefdb6d03b1b2b)
+
+changes:
+* doxygen-awesome-interactive-toc.js patched for correct loading
