+++ scripts/docs/footer.html
@@ -51,13 +51,12 @@ $generatedby <a href="https://www.doxygen.org/index.html">Doxygen</a> $doxygenve
 
       if (isLanding) return;
 
-      draw_toc();
       highlight_code();
       styleNavButtons();
     }, 0);
     });
   });
 </script>
-<script type="module" src="$relpath^feedback.js"></script>
+<script type="module" src="$relpath^toc.js"></script>
 </body>
 </html>
