+++ scripts/docs/header.html
@@ -53,13 +53,11 @@
       <script type="text/javascript" src="$relpath^doxygen-awesome-fragment-copy-button.js"></script>
       <script type="text/javascript" src="$relpath^doxygen-awesome-paragraph-link.js"></script>
       <script type="text/javascript" src="$relpath^doxygen-awesome-interactive-toc.js"></script>
-      <script type="text/javascript" src="$relpath^toc.js"></script>
       <script type="text/javascript">
         DoxygenAwesomeDarkModeToggle.init()
         DoxygenAwesomeFragmentCopyButton.init()
         DoxygenAwesomeParagraphLink.init()
         DoxygenAwesomeInteractiveToc.topOffset = 250;
-        DoxygenAwesomeInteractiveToc.init()
       </script>
       <!-- Doxygen Awesome -->
     </head>
